Convert helper functions to pure pipes
