+++ web/src/app/modules/shared/pipes/formatpath/formatpath.pipe.spec.ts
@@ -1,0 +1,8 @@
+import { FormatPathPipe } from './formatpath.pipe';
+
+describe('FormatpathPipe', () => {
+  it('create an instance', () => {
+    const pipe = new FormatPathPipe();
+    expect(pipe).toBeTruthy();
+  });
+});
