+++ web/src/app/modules/shared/pipes/formatpath/formatpath.pipe.ts
@@ -1,0 +1,14 @@
+import { Pipe, PipeTransform } from '@angular/core';
+
+@Pipe({
+  name: 'formatpath',
+})
+export class FormatPathPipe implements PipeTransform {
+  transform(path: string): string {
+    if (!path.startsWith('/')) {
+      return '/' + path;
+    }
+
+    return path;
+  }
+}
