+++ web/src/app/modules/shared/pipes/relative/relative.pipe.spec.ts
@@ -1,0 +1,8 @@
+import { RelativePipe } from './relative.pipe';
+
+describe('RelativePipe', () => {
+  it('create an instance', () => {
+    const pipe = new RelativePipe();
+    expect(pipe).toBeTruthy();
+  });
+});
