+++ web/src/app/modules/sugarloaf/components/smart/navigation/navigation.component.ts
@@ -48,12 +48,4 @@ export class NavigationComponent implements OnInit, OnDestroy {
     const data = JSON.parse(message.data);
     this.behavior.next(data);
   };
-
-  formatPath(path: string): string {
-    if (!path.startsWith('/')) {
-      return '/' + path;
-    }
-
-    return path;
-  }
 }
